add damage system

+++ Assets/Scripts/Components/EnemyHPComponent.cs
@@ -2,5 +2,5 @@ using Unity.Entities;
 
 public struct EnemyHPComponent : IComponentData
 {
-    public float hp;
+    public int hp;
 }
